Fix Portal colors (#141)
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

Fixes # (issue)

## Type of change

The commit type is used by semantic-release to determine the version to
be released to NPM.
Put the type of change in the PR title and the name of the commit that
lands on main.
Here are the available types:

- [ ] feat: A new feature. Triggers a MINOR release
- [x] fix: A bug fix. Triggers a PATCH release
- [ ] docs: Documentation only changes
- [ ] style: Changes that do not affect the meaning of the code
(white-space, formatting, missing semi-colons, etc)
- [ ] refactor: A code change that neither fixes a bug nor adds a
feature
- [ ] perf: A code change that improves performance. Triggers a PATCH
release
- [ ] test: Adding missing or correcting existing tests
- [ ] chore: Changes to the build process or auxiliary tools and
libraries such as documentation generation
- [ ] BREAKING CHANGE: Triggers a MAJOR release.